requirements.txt
streamlit~=1.33.0
psycopg2~=2.9.9
uuid~=1.30
ultralytics~=8.2.2
pillow~=10.2.0
pandas~=2.2.1
numpy~=1.26.3
opencv-python~=4.9.0.80
torch~=2.3.0+cu121